moved kommLIst function to company jsx. updated KommLIst to use a drop down
